+++ src/VaultSharp/V1/SecretsEngines/PKI/PKISecretsEngineProvider.cs
@@ -1,4 +1,5 @@
-﻿using System.Net.Http;
+﻿using System.Collections.Generic;
+using System.Net.Http;
 using System.Threading.Tasks;
 using VaultSharp.Core;
 using VaultSharp.V1.Commons;
@@ -66,5 +67,31 @@ namespace VaultSharp.V1.SecretsEngines.PKI
             };
         }
 
+        public async Task<Secret<CertificateData>> ReadCertificateAsync(string serialNumber, string pkiBackendMountPoint = null)
+        {
+            Checker.NotNull(serialNumber, "serialNumber");
+
+            var certificateDataSecret = await _polymath.MakeVaultApiRequest<Secret<CertificateData>>(pkiBackendMountPoint ?? _polymath.VaultClientSettings.SecretsEngineMountPoints.PKI, "/cert/" + serialNumber , HttpMethod.Get, unauthenticated: true).ConfigureAwait(_polymath.VaultClientSettings.ContinueAsyncTasksOnCapturedContext);
+            
+            // Populate properties not set by vault API
+            if (string.IsNullOrEmpty(certificateDataSecret.Data.SerialNumber))
+            {
+                certificateDataSecret.Data.SerialNumber = serialNumber;
+            }
+
+            if (certificateDataSecret.Data.CertificateFormat == CertificateFormat.None)
+            {
+                certificateDataSecret.Data.CertificateFormat = CertificateFormat.pem;
+            }
+
+            return certificateDataSecret;
+        }
+
+        public async Task<Secret<CertificateKeys>> ListCertificatesAsync(string pkiBackendMountPoint = null)
+        {
+            var result = await _polymath.MakeVaultApiRequest<Secret<CertificateKeys>>(pkiBackendMountPoint ?? _polymath.VaultClientSettings.SecretsEngineMountPoints.PKI, "/certs?list=true", HttpMethod.Get).ConfigureAwait(_polymath.VaultClientSettings.ContinueAsyncTasksOnCapturedContext);
+
+            return result;
+        }
     }
 }